- register page restructure
