+++ src/pages/register/RegisterContainer.tsx
@@ -6,8 +6,8 @@ import { RootState } from '../../common/redux/types'
 import { errorToast, successToast } from '../../common/utils/Toast'
 import { ModalWrapper } from '../../common/components/Wrappers/ModalWrapper'
 import { AgentRoles } from '../../types/models'
-import { Banner } from './Banner'
-import { TextBlock } from './TextBlock'
+import { Banner } from './components/Banner'
+import { TextBlock } from './components/TextBlock'
 import { deviceWidth } from '../../lib/commonData'
 import MediaQuery from 'react-responsive'
 import { Button, ButtonTypes } from '../../common/components/Form/Buttons'
