+++ src/routes/index.tsx
@@ -5,9 +5,9 @@ import { EntitiesContainerConnected } from '../modules/Entities/Entities.contain
 import { EntitiesSelectConnected } from '../modules/Entities/Entities.container.select'
 import { ProjectCreateConnected } from '../components/project/ProjectCreate'
 import { contentType } from '../types/models'
-import { ProjectForm } from '../components/project/ProjectForm'
+import { ProjectForm } from '../pages/json'
 import { Fragment } from 'react'
-import { RegisterConnected } from '../components/register/RegisterContainer'
+import { RegisterConnected } from '../pages/register/RegisterContainer'
 import { NotFound } from '../components/public/NotFound'
 import { Spinner } from '../common/components/Spinner'
 import { ProjectSignAndCreateConnected } from '../components/project/curation/ProjectSignAndCreate'
